Add default directory to payjoin-cli
Currently Cli looks for config file in
any directory where  it is run from .
This commit changes that , it first looks in .payjoin
depending on the command run. if its receive , it
looks for config in .payjoin/receive and if it's
send , it looks for config in ./payjoin/send.
if no configs are there it checks the current directory.